pyproject.toml
[tool.poetry]
name = "casa"
version = "0.1.0"
description = ""
authors = ["aos"]

[tool.poetry.scripts]
casa = "casa:main"

[tool.poetry.dependencies]
python = "^3.10"
python-kasa = "^0.5.0"
PyYAML = "^6.0"
APScheduler = "^3.9.1"
starlette = "^0.21.0"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
flake8 = "^5.0.4"
mypy = "=0.961"
pytest = "^7.1.3"
isort = "^5.10.1"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["/nix/store/*", "result/*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
